Fixed log file generation

+++ DeadlineManager.cpp
@@ -9,12 +9,12 @@ using namespace std::chrono;
 //
 DeadlineManager::DeadlineManager()
 {
-    deadline.setup_deadline(static_cast<int>(DeadlineType::keep_alive), 500ms);
-    deadline.setup_deadline(static_cast<int>(DeadlineType::logging), 50ms);
-    deadline.setup_always_contact(static_cast<int>(DeadlineType::keep_alive));
+    deadline.setup_deadline(DeadlineType::keep_alive, 500ms);
+    deadline.setup_deadline(DeadlineType::logging, 50ms);
+    deadline.setup_always_contact(DeadlineType::keep_alive);
 }
 
 bool DeadlineManager::expired(const DeadlineType k) const
 {
-    return deadline.expired(static_cast<int>(k));
+    return deadline.expired(k);
 }
